chore: moving attendance to within NoteEditor
